Support file uploads larger than 4 GiB

+++ crates/common/src/config/groupware.rs
@@ -55,7 +55,7 @@ pub struct GroupwareConfig {
     pub default_addressbook_display_name: Option<String>,
 
     // File storage settings
-    pub max_file_size: usize,
+    pub max_file_size: u64,
 
     // Sharing settings
     pub max_shares_per_item: usize,
@@ -120,7 +120,7 @@ impl GroupwareConfig {
                 RegistryVCardVersion::V3 => VCardVersion::V3_0,
                 RegistryVCardVersion::V4 => VCardVersion::V4_0,
             },
-            max_file_size: file.max_size as usize,
+            max_file_size: file.max_size,
             alarms_enabled: alarm.enable,
             alarms_minimum_interval: alarm.min_trigger_interval.into_inner().as_secs() as i64,
             alarms_allow_external_recipients: alarm.allow_external_rcpts,
